+++ libraries/hasm/src/hasm/SymbolTable.cpp
@@ -8,14 +8,14 @@ SymbolTable::SymbolTable( std::unordered_map< std::string, Hack::word > symbols 
 {
 }
 
-void SymbolTable::addEntry( const std::string& symbol, const Hack::word address )
+bool SymbolTable::addEntry( const std::string& symbol, const Hack::word address )
 {
-    m_table.emplace( symbol, address );
+    return m_table.emplace( symbol, address ).second;
 }
 
 bool SymbolTable::contains( const std::string& symbol ) const
 {
-    return m_table.find( symbol ) != m_table.end();
+    return m_table.contains( symbol );
 }
 
 std::optional< Hack::word > SymbolTable::getAddress( const std::string& symbol ) const